pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pymarketstore"
version = "0.18"
description = "Marketstore python driver"
readme = "README.md"
license = "Apache-2.0"
authors = [
    { name = "Alpaca", email = "oss@alpaca.markets" }
]
keywords = ["database", "pandas", "financial", "timeseries"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Database",
    "Topic :: Office/Business :: Financial",
]
requires-python = ">=3.9"
dependencies = [
    "click>=8.1.0",
    "grpcio>=1.60.0",
    "msgpack>=1.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "protobuf>=5.26.0",
    "requests>=2.31.0",
    "tabulate>=0.9.0",
    "urllib3>=2.0.0",
    "websocket-client>=1.6.0",
    "websockets>=12.0",
]

[project.scripts]
bench = "benchmarks:main"
pymkts = "pymarketstore.cli:main"

[project.optional-dependencies]
dev = [
    "grpcio-tools>=1.60.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "coverage>=7.4.0",
    "ruff>=0.2.0",
]

[project.urls]
Homepage = "https://github.com/alpacahq/pymarketstore"
Repository = "https://github.com/alpacahq/pymarketstore"
Issues = "https://github.com/alpacahq/pymarketstore/issues"

[tool.hatch.build.targets.sdist]
exclude = [
    "/.circleci",
    "/tests",
    "/docs",
]

[tool.hatch.build.targets.wheel]
packages = ["pymarketstore"]

[tool.pytest.ini_options]
addopts = "--verbose --cov=pymarketstore --cov-report=term-missing"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["pymarketstore"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]

[tool.ruff]
line-length = 90
target-version = "py312"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.format]
line-ending = "lf"

[tool.ruff.lint.isort]
case-sensitive = true
lines-between-types = 1
lines-after-imports = 2
section-order = [
    "future",
    "standard-library",
    "third-party",
    "first-party",
    "local-folder",
    "tests",
]

[tool.ruff.lint.isort.sections]
tests = ["tests"]

[dependency-groups]
dev = [
    "pyqt6>=6.10.2",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=7.0.0",
    "pytest-xdist>=3.8.0",
    "qtconsole>=5.7.1",
    "ruff>=0.15.1",
]
